% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AclustsCCASampler.R
\docType{class}
\name{AclustsCCASampler-class}
\alias{AclustsCCASampler-class}
\title{Build AclustsCCA sampler}
\value{

}
\description{
class AclustsCCASampler, inherited from mmctSamplerGeneric
For more information: \url{https://cran.r-project.org/web/packages/simctest/vignettes/simctest-mmctest-intro.pdf}
}
