% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AclustsCCASampler.cancors.R
\docType{class}
\name{AclustsCCASampler.cancors-class}
\alias{AclustsCCASampler.cancors-class}
\title{Build AclustsCCA sampler based on canonical correlation}
\value{

}
\description{
class AclustsCCASampler.cancors, inherited from mmctSamplerGeneric
For more information: \url{https://cran.r-project.org/web/packages/simctest/vignettes/simctest-mmctest-intro.pdf}
}
