% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AclustsCCASampler.tailprob.R
\docType{class}
\name{AclustsCCASampler.tailprob-class}
\alias{AclustsCCASampler.tailprob-class}
\title{Build AclustsCCA sampler based on canonical correlation}
\value{

}
\description{
class AclustsCCASampler.tailprob, inherited from mmctSamplerGeneric
For more information: \url{https://cran.r-project.org/web/packages/simctest/vignettes/simctest-mmctest-intro.pdf}
}
