% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmctest.R
\name{mmctest}
\alias{mmctest}
\alias{hBH}
\alias{hPC}
\alias{hBonferroni}
\alias{tau}
\alias{hStepUp}
\alias{hStepDown}
\alias{pEstimate}
\alias{cont,mmctestres-method}
\alias{rejProb}
\alias{confidenceLimits}
\alias{testResult}
\title{Functions associated with mmctest}
\usage{
hBH(p, threshold)

hPC(p, threshold)

hBonferroni(p, threshold)

tau(p, threshold)

hStepUp(p, threshold)

hStepDown(p, threshold)

\S4method{cont}{mmctestres}(
  data,
  steps = list(maxit = 0, maxnum = 0, undecided = 0, elapsedsec = 0, maxB = 0)
)

pEstimate(obj)

rejProb(obj)

confidenceLimits(obj)

testResult(obj)

mmctest(epsilon = 0.01, threshold = 0.1, r = 10000, h, thompson = F, R = 1000)
}
\description{
Read \href{https://cran.r-project.org/web/packages/simctest/index.html}{simctest R package} for more detailed explanation.
}
