% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.residual.R
\name{partial.residual}
\alias{partial.residual}
\title{partial.residual}
\usage{
partial.residual(data, Z, nthread)
}
\arguments{
\item{data}{: A data matrix with \eqn{n} rows.}

\item{Z}{: A \eqn{n} by \eqn{r} confounder data matrix, where \eqn{n} is sample size and \eqn{r} is number of potential confounders}

\item{nthread}{: A number of threads to parallelize regression.}
}
\value{
Returns a data matrix with same dimension as \code{data}
}
\description{
Regress \code{data} onto \code{Z}
}
\details{
Title
}
