% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AclustsCCA.R
\name{plot_AclustsCCA}
\alias{plot_AclustsCCA}
\title{Creates a heatmap of significant CpG sites/Genes and exposures}
\usage{
plot_AclustsCCA(obj, annot, abs = TRUE)
}
\arguments{
\item{obj}{Output of \strong{\code{AclustsCCA}}}

\item{annot}{A preloaded annotation file that includes columns the below name. For more information, \href{https://bioconductor.org/packages/release/data/annotation/html/IlluminaHumanMethylation450kanno.ilmn12.hg19.html}{IlluminaHumanMethylation450kanno.ilmn12.hg19}.
\itemize{
\item IlmnID: a character vector of IlmnID
\item CHR: a numeric vector of chromosome containing the CpG
\item Coordinate_37: a numeric vector of chromosomal coordinates of the CpG
\item Islands_Name: a character vector of chromosomal coordinates of the CpG Island
\item Relation_to_Island: a character vector of the location of the CpG relative to the CpG island
\item UCSC_RefGene_Name: a character vector of Target gene name(s)
}}

\item{abs}{A logical flag for showing absolute value of loadings or not}
}
\value{
A heatmap of significant CpG sites/Genes and exposures
where columns are CpG sites/Genes and rows are exposures.
Selected exposures are colored otherwise remains white.
}
\description{
This function creates a heatmap of significant CpG sites/Genes and exposures
}
