% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simctest.R
\docType{class}
\name{sampalg-class}
\alias{sampalg-class}
\title{Sequential implementation of Monte Carlo tests}
\arguments{
\item{gensample:}{function that performs one sampling step. Returns 0 (sampled test statistic does not exceed the observation) or 1 (sampled test static exceeds the observation).}

\item{level:}{level passed to \code{\link{getalgonthefly}}}

\item{epsilon:}{error bound epsilon passed to \code{\link{getalgonthefly}}}

\item{maxsteps:}{maximal number of steps to take}
}
\description{
Wrapper function for convenient use of the sequential implementation of the Monte Carlo test.
Downloaded from simctest_2.6.tar.gz. This function is loaded to run \code{\link{mmctest}}.


### INPUT
}
