% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.data.R
\docType{data}
\name{sample.data}
\alias{sample.data}
\title{Sample Data for AclustsCCA R Package}
\format{
A \code{list} of length 10, which are:
\describe{
\item{DATA.X}{A \eqn{n=500} by \eqn{p} matrix where \eqn{n} is a number of subjects and \eqn{p} is a number of exposures.}
\item{DATA.Y}{A \eqn{n=500} by \eqn{q} matrix where \eqn{n} is a number of subjects and \eqn{q} is a number of CpG sites.}
\item{DATA.Z}{A \eqn{n=500} by 2 matrix where \eqn{n} is a number of subjects and 2 is a number of confounders.}
\item{TRUE.CpGs}{A vector of 28 true CpG sites that are associated with exposures}
\item{TRUE.Exposures}{A vector of 5 true exposures that are associated with outcome}
\item{TRUE.ALPHA}{A vector of length \eqn{p} which represents true loading vector \eqn{alpha}}
\item{TRUE.BETA}{A vector of length \eqn{q} which represents true loading vector \eqn{beta}}
\item{SIGMA.XX}{A \eqn{p} by \eqn{p} exposure correlation matrix used to generated data}
\item{SIGMA.YY}{A \eqn{q} by \eqn{q} outcome correlation matrix used to generated data}
\item{SIGMA.XY}{A \eqn{p} by \eqn{q} correlation matrix of exposure and outcome used to generated data}
\item{clusters.list}{A list of 9 clusters with CpG sites obtained using A-clustering, each item is a cluster. Cluster 2,5,6,7,9 are the true DMRs.}
}
}
\usage{
sample.data
}
\description{
Generates synthetic data with p=20 exposures and q=100 DNA methylation probe data in chromosome 7.
A correlation structure of exposures are defined as a block diagonal
where each block is size of five with the first order autoregressive correlation structure of correlation 0.7.
A correlation structure of outcomes are defined as a block diagonal
where each block is size of its clusters defined by bumphunter with the first order autoregressive correlation structure of correlation 0.9.
The data is generated from multivariate normal distribution.

A five exposures (exposures 1,2,3,9, and 10) are associated with a five DMR regions (each of size 2,3,4,7,and 12).
}
\keyword{datasets}
