% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCCASampler.R
\docType{class}
\name{SCCASampler-class}
\alias{SCCASampler-class}
\title{Build SCCA.SAMPLER}
\arguments{
\item{gen:}{function that performs one sampling step. Returns 0 (sampled test statistic does not exceed the observation) or 1 (sampled test static exceeds the observation)}

\item{method:}{which method to use for stopping}
}
\value{
\code{mctest}, \code{mctest.RL} and \code{mctest.simctest} all return an object of class type \code{mctestres}, which has a print function (\code{print.mctestres}).
An object of class \code{mctestres} is a list with the following components: step (total batched number of samples drawn), decision.interval (interval for the p-value), decision (expressing significance), est.p (an estimate of the p-value) and realn (the actual number of samples taken without batching).
}
\description{
Build SCCA.SAMPLER
}
\examples{
Example use
}
