% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseCCA.R
\name{SparseCCA}
\alias{SparseCCA}
\title{Implement SparseCCA}
\usage{
SparseCCA(
  X,
  Y,
  Xmethod = c("lasso", "alasso", "gglasso", "SGL"),
  Ymethod = c("lasso", "alasso", "SGL", "OLS"),
  init.method = "SVD",
  X.groupidx = NULL,
  Y.groupidx = NULL,
  standardize = T,
  max.iter = 50,
  conv = 5 * 10^-2
)
}
\arguments{
\item{X}{: \eqn{n} by \eqn{p} exposure data matrix, where \eqn{n} is sample size and \eqn{p} is number of exposures.}

\item{Y}{: \eqn{n} by \eqn{q} outcome data matrix, where \eqn{n} is sample size and \eqn{q} is number of outcomes.}

\item{Xmethod}{: penalty function for the exposure, i.e. penalty function when regressing Y onto X. Possible values are:
- "lasso": Lasso
- "alasso": Adaptive Lasso
- "gglasso": Group Lasso
- "SGL": Sparse Group Lasso}

\item{Ymethod}{: penalty function for the outcome, i.e. penalty function when regressing X onto Y. Possible values are:
- "lasso": Lasso
- "alasso": Adaptive Lasso
- "gglasso": Group Lasso
- "SGL": Sparse Group Lasso
- "OLS": Ordinary Least Square}

\item{init.method}{: "Possible two values are:
- "lasso": When high-dimensional exposure data (i.e. n<p)
- "OLS": Ordinary Least Square
- "SVD": Singular Value Decomposition}

\item{X.groupidx}{: A vector of length \eqn{p} that indicates grouping structure of exposure `X`}

\item{Y.groupidx}{: A vector of length \eqn{q} that indicates grouping structure of outcome `Y`}

\item{standardize}{: A standardization (center and scale) of exposure `X` and outcome `Y`}

\item{max.iter}{: A maximum number of iterations}

\item{conv}{: A tolerance value for convergence \eqn{epsilon}}
}
\value{
The function returns a list of 6 objects according to the following order:
  - ALPHA          : An estimated canonical vector of length \eqn{p} corresponding to the exposure data \eqn{X}
  - BETA           : An estimated canonical vector of length \eqn{q} corresponding to the outcome data \eqn{Y}
  - U_ALL          : An estimated canonical variates \eqn{X\alpha} corresponding to the exposure data \eqn{X} which is a vector of length \eqn{n}
  - V_ALL          : An estimated canonical variates \eqn{Y\beta} corresponding to the outcome data \eqn{Y} which is a vector of length \eqn{n}
  - lamdbaA        : A numeric value of the sparsity parameter lambdaA
  - lamdbaB        : A numeric value of the sparsity parameter lambdaB
  - it             : Number of iterations until convergence
}
\description{
Implement an iterative penalized least squares approach to sparse canonical correlation analysis (SparseCCA)
with various penalty functions. Modified Wilms, Ines, and Christophe Croux. "Robust sparse canonical correlation analysis." BMC systems biology 10.1 (2016): 1-13.
The original code is accessible https://sites.google.com/view/iwilms/software?authuser=0


### INPUT
}
\examples{
data.list <- generate.data(n=500)
DATA.X <- data.list$DATA.X
DATA.Y <- data.list$DATA.Y
SparseCCA.result <- SparseCCA(X=DATA.X,Y=DATA.Y,Xmethod="SGL",Ymethod="lasso",X.groupidx=c(rep(1,4),rep(2,3),rep(3,2),rep(1,3)))
str(SparseCCA.result)
projection.error(data.list$TRUE.ALPHA,SparseCCA.result$ALPHA)
projection.error(data.list$TRUE.BETA,SparseCCA.result$BETA)


}
