% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Aclustering.R
\name{summary.Aclustering}
\alias{summary.Aclustering}
\title{Summarize the A-clustering result}
\usage{
\method{summary}{Aclustering}(clusters.list, annot, digits = 2)
}
\arguments{
\item{clusters.list}{: A list of clusters with CpG sites, each item is a cluster that contains a set of probes}

\item{annot}{: A preloaded annotation file that includes columns the below name. For more information, \url{https://support.illumina.com/content/dam/illumina-support/documents/downloads/productfiles/methylationepic/infinium-methylationepic-manifest-column-headings.pdf}.
\itemize{
\item IlmnID: The IlmnID
\item Coordinate_37: Chromosomal coordinates of the CpG
\item Islands_Name: Chromosomal coordinates of the CpG Island
\item Relation_to_Island: The location of the CpG relative to the CpG island
\item UCSC_RefGene_Name: Target gene name(s)
}}

\item{digits}{: Show rounded values to the specified number of decimal places}
}
\value{
The function returns a table of summary characteristics:
\itemize{
\item Total number of sites used for the analysis
\item Number of singletons (a cluster with one CpG site)
\item Number of clusters (a cluster with at least two sites)
\item Number of CpG sites that are clustered
\item Number of CpG sites in cluster (min, median, max)
\item Base-pair distance between extremes (min, median, max)
\item (1) Clusters associated with a single refernece gene
\item (2) Clusters associated with a single CpG Island
\item (3-1) Clusters associated with a CpG resort : at least one resort regions
\item (3-2) Clusters associated with a CpG resort : at least two resort regions
\item (3-3) Clusters associated with a CpG resort : one resort regions
\item (3-4) Clusters associated with a CpG resort : two resort regions
\item (3-5) Clusters associated with a CpG resort : three resort regions
}
}
\description{
Implement an iterative penalized least squares approach to sparse canonical correlation analysis (SparseCCA)
with various penalty functions
\subsection{INPUT}{
}
}
\examples{
library("IlluminaHumanMethylation450kanno.ilmn12.hg19")
annot <- minfi::getAnnotation(IlluminaHumanMethylation450kanno.ilmn12.hg19)
annot <- data.table(data.frame(annot))
annot <- annot[!chr\%in\%c("chrY","chrX"),]
annot <- annot[,list("IlmnID"=Name,"Coordinate_37"=pos,"CHR"=as.numeric(gsub("chr", "", chr)),Islands_Name,Relation_to_Island,UCSC_RefGene_Name)]
annot <- annot[order(CHR,Coordinate_37),]

clusters.list <- assign.to.clusters(betas.7, annot)
summary.Aclustering(clusters.list=clusters.list,annot=annot,digits=2)


}
