% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_Aclustering.R
\name{summary_Aclustering}
\alias{summary_Aclustering}
\title{Create a table that summarizes A-clustering result}
\usage{
summary_Aclustering(clusters.list, annot)
}
\arguments{
\item{clusters.list}{A list of clusters with CpG sites, each item is a cluster that contains a set of probes}

\item{annot}{A preloaded annotation file that includes columns the below name. For more information, \href{https://bioconductor.org/packages/release/data/annotation/html/IlluminaHumanMethylation450kanno.ilmn12.hg19.html}{IlluminaHumanMethylation450kanno.ilmn12.hg19}.
\itemize{
\item IlmnID: a character vector of IlmnID
\item CHR: a numeric vector of chromosome containing the CpG
\item Coordinate_37: a numeric vector of chromosomal coordinates of the CpG
\item Islands_Name: a character vector of chromosomal coordinates of the CpG Island
\item Relation_to_Island: a character vector of the location of the CpG relative to the CpG island
\item UCSC_RefGene_Name: a character vector of Target gene name(s)
}}
}
\value{
The function returns a table of summary characteristics:
\itemize{
\item Total number of sites used for the analysis
\item Number of singletons (a cluster with one CpG site)
\item Number of clusters (a cluster with at least two sites)
\item Number of CpG sites that are clustered
\item Number of CpG sites in cluster (min, median, max)
\item Base-pair distance between extremes (min, median, max)
\item (1) Clusters associated with a single refernece gene
\item (2) Clusters associated with a single CpG Island
\item (3) Number of resort regions each cluster is associated with (min, median, max)
\item (3-1) Clusters associated with a CpG resort : at least one resort regions
\item (3-2) Clusters associated with a CpG resort : at least two resort regions
}
}
\description{
This function creates a table that summarizes A-clustering result.
}
\examples{
data(annot)
data(sample.data)
DATA.Y <- sample.data$DATA.Y # row: subjects (n), column: CpG sites (q)

all.clusters.list <- Aclust::assign.to.clusters(betas = t(DATA.Y),annot = annot, dist.type = "spearman", method = "average", thresh.dist = 0.2,max.dist = 1000,bp.thresh.dist = 999)
summary_Aclustering(clusters.list=all.clusters.list,annot=annot)

}
