% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_AclustsCCA.R
\name{summary_AclustsCCA}
\alias{summary_AclustsCCA}
\title{Outputs summary of AclustsCCA result}
\usage{
summary_AclustsCCA(obj, annot, n.top = NULL)
}
\arguments{
\item{obj}{Output of \strong{\code{AclustsCCA}}}

\item{annot}{A preloaded annotation file that includes columns the below name. For more information, \href{https://bioconductor.org/packages/release/data/annotation/html/IlluminaHumanMethylation450kanno.ilmn12.hg19.html}{IlluminaHumanMethylation450kanno.ilmn12.hg19}.
\itemize{
\item IlmnID: a character vector of IlmnID
\item CHR: a numeric vector of chromosome containing the CpG
\item Coordinate_37: a numeric vector of chromosomal coordinates of the CpG
\item Islands_Name: a character vector of chromosomal coordinates of the CpG Island
\item Relation_to_Island: a character vector of the location of the CpG relative to the CpG island
\item UCSC_RefGene_Name: a character vector of Target gene name(s)
}}

\item{n.top}{A number of top clusters with smallest p.value to show}
}
\value{
The function returns a table of summary characteristics for each cluster:
\itemize{
\item ClustIdx: Index number of cluster corresponding to \strong{\code{clusters.list}}
\item CHR: a numeric value of chromosome containing the CpG
\item UCSC_RefGene_Name: a list of target gene name(s)
\item Islands_Name: a list of chromosomal coordinates of the CpG Island
\item Significant: whether or not that cluster is significant
\item Cancors: a numeric value of canonical correlation
\item N.Probes: a number of probes in that cluster
\item DMR.length: a length of DMR in base pairs
\item Num.Exceed: a number of exceedness in permutation test
\item Num.Perm: a number of permutation
\item P.value: P.value corresponding to this cluster
\item CpGs: CpG sites with non-zero loadings
\item Exposures: Exposures with non-zero loadings
\item CpGs.Loading: A loading vector of CpGs sites with non-zero loadings
\item Exposures.Loading:  A loading vector of exposures with non-zero loadings
}
}
\description{
This function creates summary of AclustsCCA result table that summarizes A-clustering result.
}
