% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AclustsCCA.R
\name{AclustsCCA}
\alias{AclustsCCA}
\title{Implement AclustsCCA}
\usage{
AclustsCCA(
  clusters.list = NULL,
  X,
  Y,
  Z = NULL,
  X.resid = NULL,
  Y.resid = NULL,
  annot = NULL,
  dist.type = "spearman",
  Aclust.method = "average",
  thresh.dist = 0.2,
  max.dist = 1000,
  bp.thresh.dist = 999,
  Xmethod = "lasso",
  Ymethod = "OLS",
  standardize = T,
  X.groupidx = NULL,
  init.method = "SVD",
  max.iter = 100,
  conv = 10^-2,
  maxnum = NULL,
  maxB = 10000,
  FDR.thresh = 0.05,
  h = hBH,
  permute = T,
  nthread = 2,
  test.stat = c("cancors", "tailprob")
)
}
\arguments{
\item{clusters.list}{A list of clusters with CpG sites obtained using A-clustering, each item is a cluster that contains a set of probes. A-clustering is implemented if \code{NULL} or can be provided by users.}

\item{X}{\eqn{n} by \eqn{p} exposure data matrix, where \eqn{n} is sample size and \eqn{p} is number of exposures.}

\item{Y}{\eqn{n} by \eqn{q} outcome data matrix, where \eqn{n} is sample size and \eqn{q} is number of outcomes.}

\item{Z}{\eqn{n} by \eqn{r} confounder data matrix, where \eqn{n} is sample size and \eqn{r} is number of confounders. If \code{NULL}, partial residuals are used for SparseCCA analysis.}

\item{annot}{A preloaded annotation file that includes columns "IlmnID", "Coordinate_37", "Islands_Name", "Relation_to_Island", "UCSC_RefGene_Name". Only needed if \strong{\code{clusters.list}} is \code{NULL}.}

\item{dist.type}{A type of similarity distance function. Options are "spearman" (default), "pearson" (correlation measures) or "euclid".}

\item{Aclust.method}{A type of clustering function. Options are "single", "complete" or "average" (default).}

\item{thresh.dist}{A similarity distance threshold. Two neighboring clusters are merged to a single cluster if the similarity distance between them is above dist.thresh. The default is 0.2}

\item{max.dist}{Optional maximum length between neighboring variables permitting to cluster them together. The default is 1000.}

\item{bp.thresh.dist}{A distance in chromosomal location. Any set of methylation sites within an interval smaller or equal to \strong{\code{bp.dist}} will be potentially merged, depending on the similarity between sites at the ends of the interval. The default is 999.}

\item{Xmethod}{A penalty function for the exposure, i.e. penalty function when regressing Y onto X. Options are "lasso", "alasso","gglasso", and "SGL" (default).}

\item{Ymethod}{A penalty function for the outcome, i.e. penalty function when regressing X onto Y. Options are "lasso", "alasso","gglasso", "SGL", and "OLS" (default).}

\item{standardize}{A logical flag for exposure \eqn{X} and outcome \eqn{Y} standardization, prior to fitting the model.}

\item{X.groupidx}{A vector of length \eqn{p} that indicates grouping structure of exposure \eqn{X}.}

\item{init.method}{Initialization method. Options are "lasso", "OLS", and "SVD" (default).}

\item{max.iter}{A maximum number of iterations of SparseCCA. The default is 100.}

\item{conv}{A tolerance value for convergence \eqn{epsilon} of SparseCCA. The default is 10e-2.}

\item{maxnum}{A maximal total number of permutations across all the clusters.}

\item{maxB}{A maximal number of permutations for a single cluster.}

\item{FDR.thresh}{False discovery rate (FDR) threshold. The default is 0.05.}

\item{permute}{A logical flag for whether to run permutation test or not.}

\item{nthread}{A number of threads to parallelize permutation test and implementation of SparseCCA across all the clusters.}

\item{test.stat}{A test statistic for permutation test. Options are canonical correlations ("cancors") or tail probability ("tailprob").}

\item{permute.tmp.filepath}{A file path to save intermittent permutation results.}
}
\value{
The function returns a list of 6 objects according to the following order:
\itemize{
\item clusters.list          : A list of clusters with CpG sites obtained using A-clustering, each item is a cluster that contains a set of probes. If A-clustering is not implemented inside AclustsCCA, return \code{NA}.
\item ALPHA.observed         : A list of estimated canonical vector of length \eqn{p} corresponding to the exposure data \eqn{X} for each cluster.
\item BETA.observed          : A list of estimated canonical vector of length \eqn{q} corresponding to the outcome data \eqn{Y} for each cluster.
\item cancors.observed       : A vector of estimated canonical correlation for each cluster.
\item permutation.result     : A \code{mmctest} object that contains permutation results.
\item settings               : A settings used for the analysis.
}
}
\description{
Implement an iterative penalized least squares approach to
sparse canonical correlation analysis (SparseCCA)
with various penalty functions.
}
\examples{



}
