% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmctest.R
\name{hBH}
\alias{hBH}
\title{Sequential implementation of the Monte Carlo test with p-value buckets.}
\usage{
hBH(p, threshold)
}
\description{
Class which creates an object of type "mmctestres". Objects can be created by calls of the form.
}
\examples{
  fun <- function(ind,n,data) sapply(1:length(ind), function(i) sum(runif(n[i])<=data[ind[i]]));
  i <- mmctSampler(fun,num=500,data=runif(500));
  a <- mmctest(h=hBH);
  a <- run(a, i, maxsteps=list(maxnum=1000000,undecided=10));
}
