% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseCCA.R
\name{SparseCCA}
\alias{SparseCCA}
\title{Implement SparseCCA}
\usage{
SparseCCA(
  X,
  Y,
  Xmethod = c("lasso", "alasso", "gglasso", "SGL"),
  Ymethod = c("lasso", "alasso", "OLS"),
  init.method = "SVD",
  X.groupidx = NULL,
  standardize = T,
  max.iter = 50,
  conv = 5 * 10^-2
)
}
\arguments{
\item{X}{: \eqn{n} by \eqn{p} exposure data matrix, where \eqn{n} is sample size and \eqn{p} is number of exposures.}

\item{Y}{: \eqn{n} by \eqn{q} outcome data matrix, where \eqn{n} is sample size and \eqn{q} is number of outcomes.}

\item{Xmethod}{: penalty function for the exposure, i.e. penalty function when regressing Y onto X. Possible values are:
\itemize{
\item "lasso": Lasso
\item "alasso": Adaptive Lasso
\item "gglasso": Group Lasso
\item "SGL": Sparse Group Lasso
}}

\item{Ymethod}{: penalty function for the outcome, i.e. penalty function when regressing X onto Y. Possible values are:
\itemize{
\item "lasso": Lasso
\item "alasso": Adaptive Lasso
\item "OLS": Ordinary Least Square
}}

\item{init.method}{: "Possible two values are:
\itemize{
\item "lasso": When high-dimensional exposure data (i.e. n<p)
\item "OLS": Ordinary Least Square
\item "SVD": Singular Value Decomposition
}}

\item{X.groupidx}{: A vector of length \eqn{p} that indicates grouping structure of exposure \code{X}}

\item{standardize}{: A logical flag for exposure \eqn{X} and outcome \eqn{Y} standardization, prior to fitting the model.}

\item{max.iter}{: A maximum number of iterations}

\item{conv}{: A tolerance value for convergence \eqn{epsilon}}
}
\value{
The function returns a list of 6 objects according to the following order:
\itemize{
\item ALPHA          : An estimated canonical vector of length \eqn{p} corresponding to the exposure data \eqn{X}
\item BETA           : An estimated canonical vector of length \eqn{q} corresponding to the outcome data \eqn{Y}
\item U_ALL          : An estimated canonical variates \eqn{X\alpha} corresponding to the exposure data \eqn{X} which is a vector of length \eqn{n}
\item V_ALL          : An estimated canonical variates \eqn{Y\beta} corresponding to the outcome data \eqn{Y} which is a vector of length \eqn{n}
\item lamdbaA        : A numeric value of the sparsity parameter lambdaA
\item lamdbaB        : A numeric value of the sparsity parameter lambdaB
\item it             : Number of iterations until convergence
}
}
\description{
Implement an iterative penalized least squares approach to sparse canonical correlation analysis (SparseCCA)
with various penalty functions. Modified Wilms, Ines, and Christophe Croux. "Robust sparse canonical correlation analysis." BMC systems biology 10.1 (2016): 1-13.
The original code is accessible https://sites.google.com/view/iwilms/software?authuser=0
\subsection{INPUT}{
}
}
\examples{
data.list <- generate.data(n=500)
DATA.X <- data.list$DATA.X
DATA.Y <- data.list$DATA.Y
SparseCCA.result <- SparseCCA(X=DATA.X,Y=DATA.Y,Xmethod="SGL",Ymethod="lasso",X.groupidx=c(rep(1,4),rep(2,3),rep(3,2),rep(1,3)))
str(SparseCCA.result)
projection.error(data.list$TRUE.ALPHA,SparseCCA.result$ALPHA)
projection.error(data.list$TRUE.BETA,SparseCCA.result$BETA)


}
